requirements.txt
im-data-manager-job-decoder == 2.1.0
munch == 4.0.0
pyyaml == 6.0.1
